+++ beagle/beagled/Lucene.Net/Store/RAMInputStream.cs
@@ -1,9 +1,10 @@
 /*
- * Copyright 2004 The Apache Software Foundation
- * 
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
  * 
  * http://www.apache.org/licenses/LICENSE-2.0
  * 
@@ -22,13 +23,13 @@ namespace Lucene.Net.Store
 	/// <summary> A memory-resident {@link IndexInput} implementation.
 	/// 
 	/// </summary>
-	/// <version>  $Id: RAMInputStream.cs,v 1.3 2006/10/02 17:09:10 joeshaw Exp $
+	/// <version>  $Id: RAMInputStream.java 150537 2004-09-28 20:45:26Z cutting $
 	/// </version>
 	
 	class RAMInputStream : BufferedIndexInput, System.ICloneable
 	{
 		private RAMFile file;
-		private int pointer = 0;
+		private long pointer = 0;
 		private long length;
 		
 		public RAMInputStream(RAMFile f)
@@ -40,11 +41,11 @@ namespace Lucene.Net.Store
 		public override void  ReadInternal(byte[] dest, int destOffset, int len)
 		{
 			int remainder = len;
-			int start = pointer;
+			long start = pointer;
 			while (remainder != 0)
 			{
-				int bufferNumber = start / BUFFER_SIZE;
-				int bufferOffset = start % BUFFER_SIZE;
+				int bufferNumber = (int) (start / BUFFER_SIZE);
+				int bufferOffset = (int) (start % BUFFER_SIZE);
 				int bytesInBuffer = BUFFER_SIZE - bufferOffset;
 				int bytesToCopy = bytesInBuffer >= remainder?remainder:bytesInBuffer;
 				byte[] buffer = (byte[]) file.buffers[bufferNumber];
@@ -62,7 +63,7 @@ namespace Lucene.Net.Store
 		
 		public override void  SeekInternal(long pos)
 		{
-			pointer = (int) pos;
+			pointer = pos;
 		}
 		
 		public override long Length()
